backend/requirements.txt
django==5.0.8
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-dotenv==1.0.1
